requirements.txt
pydantic
pandas
openpyxl
pdfplumber
PyPDF2
google-genai
python-dotenv
